Split out table_impl & reducer_impl into their own modules
